refactor: 로컬 커스텀 훅 관련있는 컴포넌트 내의 별도 파일로 분리
* 뷰와 상태관련 로직을 분리해서 직관적이고, 유지보수 측면에서 이점을
  가져가기 위해 별도의 파일로 분리했습니다.
* 2개 이상의 컴포넌트에서 사용되는 경우 hooks 디렉토리에 저장합니다.
* 하나의 컴포넌트에서 사용되는 경우 해당 컴포넌트 디렉토리에 결합성을
  강하게 가져가도록 위치합니다.
